Extract twiml logic from inboundCall controller to service
