--- src/services/participant.service.ts
@@ -1,5 +1,0 @@
-export async function addParticipantToConversation(
-  params: AddParticipantToConversation
-) {}
-
-export interface AddParticipantToConversation {}
